# doc/api.yaml
openapi: 3.0.0
info:
  title: React Learn Words
  description: a backend part of RSSchool-React "Learn words" task
  version: 1.0.0
servers:
  - url: /
components:
  schemas:
...
    Statistic:
      type: object
      properties:
        gameDate:
          type: string
          format: date-time
          example: "2022-09-02T09:16:03Z"
        gameType:
          type: string
          enum: [audio-challenge, sprint]
        learnedWords:
          type: number
        guessedWords:
          type: number
        totalWords:
          type: number
        maxGuessedSeries:
          type: number
        optional:
          type: object
          items:
            type: any
...
  responses:
    UnauthorizedError:
      description: Access token is missing or invalid
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
...
  /users/{id}/statistics:
    parameters:
      - name: id
        description: userId
        in: path
        required: true
        schema:
          type: string
    get:
      tags:
        - Users/Statistics
      summary: Get all statistics for a user
      description: Get all statistics for a user
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Statistic"
        401:
          $ref: '#/components/responses/UnauthorizedError'
        404:
          description: Statistics not found
    post:
      tags:
        - Users/Statistics
      summary: Create statistic record
      description: Create statistic record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              title: example
              properties:
                gameDate:
                  type: string
                  format: date-time
                  example: "2022-09-02T09:16:03Z"
                gameType:
                  type: string
                  enum: [audio-challenge, sprint]
                learnedWords:
                  type: number
                guessedWords:
                  type: number
                totalWords:
                  type: number
                maxGuessedSeries:
                  type: number
                optional:
                  type: object
                  items:
                    type: any
      responses:
        200:
          description: The statistics record has been created.
          content:
            application/json:
              schema:
                  $ref: "#/components/schemas/Statistic"
        400:
          description: Bad request
        401:
          $ref: '#/components/responses/UnauthorizedError'
  /users/{id}/statistics/{statId}:
    parameters:
      - name: id
        description: userId
        in: path
        required: true
        schema:
          type: string
      - name: statId
        description: statId
        in: path
        required: true
        schema:
          type: string
    get:
      tags:
        - Users/Statistics
      summary: Get a statistics record by id
      description: Get a statistics record by id
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Statistic"
        401:
          $ref: '#/components/responses/UnauthorizedError'
        404:
          description: Statistics record not found
    put:
      tags:
        - Users/Statistics
      summary: Update a statistics record
      description: Update a statistics record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              title: example
              properties:
                gameDate:
                  type: string
                  format: date-time
                  example: "2022-09-02T09:16:03Z"
                gameType:
                  type: string
                  enum: [audio-challenge, sprint]
                learnedWords:
                  type: number
                guessedWords:
                  type: number
                totalWords:
                  type: number
                maxGuessedSeries:
                  type: number
                optional:
                  type: object
                  items:
                    type: any
      responses:
        200:
          description: The statistics record has been updated.
          content:
            application/json:
              schema:
                  $ref: "#/components/schemas/Statistic"
        400:
          description: Bad request
        401:
          $ref: '#/components/responses/UnauthorizedError'
    delete:
      tags:
        - Users/Statistics
      summary: Delete a statistics record
      description: Delete a statistics record
      responses:
        204:
          description: The statistics record has been deleted
        401:
          $ref: '#/components/responses/UnauthorizedError'
...